Implementing ExchangeRate endpoint

+++ src/main/java/ru/cyberpunkoff/currencyexchange/dao/CurrencyDao.java
@@ -8,7 +8,7 @@ import java.util.List;
 public interface CurrencyDao {
     List<Currency> findAll() throws SQLException;
 
-    long insertCurrency(Currency currency) throws SQLException;
+    int insertCurrency(Currency currency) throws SQLException;
 
     Currency findById(int id) throws SQLException;
